Set device info from scopes on discovery
Currently only name and hardware (model) since only they have
corresponding DeviceInfo fields. The other two standard scopes that are
not yet stored are location and profile. If anyone wants them they can
be added later.

One detail is that in the spec it's described that a single scope can
have multiple values. This is not taken into consideration in this
implementation.

See section 7.3.2.2 Scopes in the ONVIF core spec:
https://www.onvif.org/specs/core/ONVIF-Core-Specification.pdf